% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD_object_functions.R
\docType{methods}
\name{c,countData-method}
\alias{c,countData-method}
\title{c method for testClass}
\usage{
\S4method{c}{countData}(x, ..., recursive = FALSE)
}
\arguments{
\item{x}{Value}

\item{...}{Value}

\item{recursive}{Value}
}
\description{
c method for testClass
}
