% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD_object_functions.R
\docType{methods}
\name{calcNormFactors}
\alias{calcNormFactors}
\title{MyHappyFunction: The my happy function}
\usage{
calcNormFactors(object, ...)
}
\arguments{
\item{object}{Value}

\item{...}{Value}
}
\description{
calcNormFactors method for testClass
}
