% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD_object_functions.R
\docType{methods}
\name{calcNormFactors.DGEList}
\alias{calcNormFactors.DGEList}
\title{calcNormFactors.DGEList method for testClass}
\usage{
\method{calcNormFactors}{DGEList}(object, method = c("TMM", "RLE",
  "upperquartile", "none"), refColumn = NULL, logratioTrim = 0.3,
  sumTrim = 0.05, doWeighting = TRUE, Acutoff = -1e+10, p = 0.75,
  ...)
}
\arguments{
\item{object}{Value}

\item{method}{Value}

\item{refColumn}{Value}

\item{logratioTrim}{Value}

\item{sumTrim}{Value}

\item{doWeighting}{Value}

\item{Acutoff}{Value}

\item{p}{Value}

\item{...}{Value}
}
\description{
calcNormFactors.DGEList method for testClass
}
