% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD_object_functions.R
\docType{class}
\name{countData-class}
\alias{countData-class}
\title{An S4 class that stores count data.}
\description{
An S4 class that stores count data.
}
\section{Slots}{

\describe{
\item{\code{data}}{Contains data.}

\item{\code{RPKM}}{Contains RPKM data.}

\item{\code{annotation}}{Contains annotation column data.}

\item{\code{cellObservables}}{Contains cellObservables data.}

\item{\code{densityFunction}}{Contains densityFunction data.}

\item{\code{estProps}}{Contains estProps data.}

\item{\code{groups}}{Contains groups data.}

\item{\code{nullPosts}}{Contains nullPosts data.}

\item{\code{orderings}}{Contains orderings data.}

\item{\code{posteriors}}{Contains posteriors data.}

\item{\code{priorModels}}{Contains priorModels data.}

\item{\code{priorType}}{Contains priorType data.}

\item{\code{priors}}{Contains priors data.}

\item{\code{replicates}}{Contains replicates data.}

\item{\code{rowObservables}}{Contains rowObservables data.}

\item{\code{sampleObservables}}{Contains sampleObservables data.}
}}

