% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dafsFilter.R
\name{dafsFilter}
\alias{dafsFilter}
\title{dafsFilter}
\usage{
dafsFilter(VEC1, PLOT)
}
\arguments{
\item{VEC1}{Vector 1.}

\item{PLOT}{Boolean, toggles plotting.}
}
\description{
This function filters the expression.
}
\details{
This function filters the expression.
}
\references{
BMC Bioinformatics, 2014, 15:92
}
\author{
AJ Vaestermark, JR Walters.
}
