% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dafsFilterWrapper.R
\name{dafsFilterWrapper}
\alias{dafsFilterWrapper}
\title{dafsFilterWrapper	Function to filter expression data within a countData object.}
\usage{
dafsFilterWrapper(cD, PLOT=TRUE)
}
\arguments{
\item{cD}{A countData object.}

\item{PLOT}{Boolean, toggles plotting.}
}
\value{
Returns an invisible filtered countData object.
}
\description{
This function filters the expression of the supplied countData object, by invoking the dafsFilter function. dafsFilter is a filtering function used to remove rows (genes) of various expression data.
}
\details{
This function filters the expression of the supplied countData object using a Data Adaptive Flag filter. The internal function (dafsFilter) uses a vector to store Kolmogorov Smirnov distance statistics, loops through cuts of the data to determine targeted K-S statistic, selects data greater than a quantile and runs Mclust on that data to determine theoretical distribution. The wrapper uses simpleFilter to determine first left-most local minima (using the Earth library).
}
\references{
BMC Bioinformatics, 2014, 15:92
}
\author{
AJ Vaestermark, JR Walters.
}
