% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD_object_functions.R
\docType{methods}
\name{densityFunction<-}
\alias{densityFunction<-}
\alias{densityFunction<-,countData-method}
\title{densityFunction method for testClass}
\usage{
densityFunction(x) <- value

\S4method{densityFunction}{countData}(x) <- value
}
\arguments{
\item{x}{Value}

\item{value}{Value}
}
\description{
densityFunction method for testClass

densityFunction method for testClass
}
