% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD_object_functions.R
\docType{methods}
\name{DGEList}
\alias{DGEList}
\title{DGEList method for testClass}
\usage{
DGEList(counts = matrix(0, 0, 0), lib.size = colSums(counts),
  norm.factors = rep(1, ncol(counts)), samples = NULL, group = NULL,
  genes = NULL, remove.zeros = FALSE)
}
\arguments{
\item{counts}{Value}

\item{lib.size}{Value}

\item{norm.factors}{Value}

\item{samples}{Value}

\item{group}{Value}

\item{genes}{Value}

\item{remove.zeros}{Value}
}
\description{
DGEList method for testClass
}
