% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD_object_functions.R
\docType{methods}
\name{dim,countData-method}
\alias{dim,countData-method}
\title{dim method for testClass}
\usage{
\S4method{dim}{countData}(x)
}
\arguments{
\item{x}{Value}
}
\description{
dim method for testClass
}
