% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD_object_functions.R
\docType{methods}
\name{dropEmptyLevels}
\alias{dropEmptyLevels}
\title{dropEmptyLevels method for testClass}
\usage{
dropEmptyLevels(x)
}
\arguments{
\item{x}{Value}
}
\description{
dropEmptyLevels method for testClass
}
