% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD_object_functions.R
\docType{methods}
\name{[,countData-method}
\alias{[,countData-method}
\title{summary method for testClass}
\usage{
\S4method{[}{countData}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{Value}

\item{i}{Value}

\item{j}{Value}

\item{...}{Value}

\item{drop}{Value}
}
\description{
summary method for testClass
}
