% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD_object_functions.R
\docType{methods}
\name{getLibsizes}
\alias{getLibsizes}
\title{getLibsizes method for testClass}
\usage{
getLibsizes(cD, data, replicates, subset = NULL,
  estimationType = c("quantile", "total", "edgeR"), quantile = 0.75,
  ...)
}
\arguments{
\item{cD}{Value}

\item{data}{Value}

\item{replicates}{Value}

\item{subset}{Value}

\item{estimationType}{Value}

\item{quantile}{Value}

\item{...}{Value}
}
\description{
getLibsizes method for testClass
}
