% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqrxFilter.R
\name{iqrxFilter}
\alias{iqrxFilter}
\title{iqrxFilter Function to filter expression data within a countData object.}
\usage{
iqrxFilter(cD, perc_cutoff=0.25, perc_cutoff2=0.25, IQR_MULTIPLIER=1.5, MEDIAN=FALSE)
}
\arguments{
\item{cD}{A countData object.}

\item{perc_cutoff}{The lower cutoff, expressed as a percentage.}

\item{perc_cutoff2}{The upper cutoff, expressed as a percentage.}

\item{IQR_MULTIPLIER}{Numeric multiplier; removes any outliers that are IQR_MULTIPLIER times the mid-50 percentile distance greater or less than the perc_cutoff and 1-perc_cutoff2 (25th and 75th percentiles, by default)}

\item{MEDIAN}{Boolean, Calculate RowMeans or RowMedians.}
}
\value{
Returns a filtered countData object.
}
\description{
This function filters the expression of the supplied countData object. iqrxFilter is a filtering function used to remove rows (genes) of various expression data.
}
\details{
This function filters the expression of the supplied countData object, based on a selected percentage cutoff and selected interquartile range multiplier. The function iqrxFilter will: (1) log-base two transform all RPKM values (obligatory); (2) remove any outliers that were 1.5 times the mid-50 percentile distance greater or less than the 75th and 25th percentiles (by default), respectively; and (3) uses mean values and instead of median values (by default).
}
\references{
Jue et al. BMC Genomics201314:150
}
\author{
AJ Vaestermark, JR Walters.
}
