% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD_object_functions.R
\docType{methods}
\name{libsizes<-}
\alias{libsizes<-}
\alias{libsizes<-,countData-method}
\alias{libsizes}
\alias{libsizes,countData-method}
\title{libsizes method for testClass}
\usage{
libsizes(x) <- value

\S4method{libsizes}{countData}(x) <- value

libsizes(x)

\S4method{libsizes}{countData}(x)
}
\arguments{
\item{x}{Value}

\item{value}{Value}
}
\description{
libsizes method for testClass

libsizes method for testClass

libsizes method for testClass

libsizes method for testClass
}
