% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExpr.R
\name{plotExpr}
\alias{plotExpr}
\title{plotExpr Function to generate a boxplot (expression) for a countData object based on the replicate data.}
\usage{
plotExpr(cD, groupings, mode_mean, treatment, LOG2, clusterby_grouping, ...)
}
\arguments{
\item{cD}{A countData object, an object for which a summary is desired.}

\item{groupings}{A grouping (annotation column), e.g. groupings="something".}

\item{mode_mean}{Boolean, Calculate RowMeans or RowMedians.}

\item{treatment}{list, indicating which treatments (columns) to be plotted.}

\item{LOG2}{Boolean, Calculate LOG2.}

\item{clusterby_grouping}{Boolean, Cluster boxplots by grouping (or treatment).}

\item{...}{Passthrough arguments to boxplot (additional arguments affecting the summary produced).}
}
\value{
Returns an invisible data frame containing values and labels.
}
\description{
This function boxplots the expression of the supplied countData object. plotExpr is a generic function used to produce boxplot summaries of various expression data. The function invokes particular plottings which depend on the selected grouping and treatment.
}
\details{
This function boxplots the expression of the supplied countData object, based on a selected annotation column and selected treatments.
}
\references{
The "doseR" package, 2018 (in press).
}
\author{
AJ Vaestermark, JR Walters.
}
