% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantFilter.R
\name{quantFilter}
\alias{quantFilter}
\title{quantFilter	Function to filter expression data of a countData object.}
\usage{
quantFilter (cD, perc_cutoff=NULL, perc_cutoff2=0.05, MEDIAN=FALSE)
}
\arguments{
\item{cD}{A countData object.}

\item{perc_cutoff}{The lower cutoff, expressed as a percentage.}

\item{perc_cutoff2}{The upper cutoff, expressed as a percentage.}

\item{MEDIAN}{Boolean, Calculate RowMeans or RowMedians.}
}
\value{
Returns a filtered countData object.
}
\description{
This function filters the expression of the supplied countData object; quantFilter is a filtering function used to remove rows (genes) of various expression data.
}
\details{
This function filters the expression of the supplied countData object, based on a selected percentage cutoff.
}
\references{
The "doseR" package, 2018 (in press).
}
\author{
AJ Vaestermark, JR Walters.
}
