% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD_object_functions.R
\docType{methods}
\name{replicates<-}
\alias{replicates<-}
\alias{replicates<-,countData-method}
\title{replicates method for testClass}
\usage{
replicates(x) <- value

\S4method{replicates}{countData}(x) <- value
}
\arguments{
\item{x}{Value}

\item{value}{Value}
}
\description{
replicates method for testClass

replicates method for testClass
}
