% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD_object_functions.R
\docType{methods}
\name{summary,countData-method}
\alias{summary,countData-method}
\title{summary method for testClass}
\usage{
\S4method{summary}{countData}(object, ...)
}
\arguments{
\item{object}{Value}

\item{...}{Value}
}
\description{
summary method for testClass
}
